Fix 32-bit warning

+++ src/import/text_matrix.h
@@ -259,7 +259,7 @@ namespace lily_of_the_valley
                     #pragma omp parallel for
                     for (int64_t row = 0; row < static_cast<int64_t>(m_matrix->size()); ++row)
                         {
-                        for (auto& cell : (*m_matrix)[row])
+                        for (auto& cell : (*m_matrix)[static_cast<size_t>(row)])
                             {
                             for (const auto& mdVal : m_mdVals.value())
                                 {
